Squash all Alembic migrations in one #28
