+++ backend/app/alembic/script.py.mako
@@ -5,8 +5,10 @@ Revises: ${down_revision | comma,n}
 Create Date: ${create_date}
 
 """
-from alembic import op
 import sqlalchemy as sa
+
+from alembic import op
+
 ${imports if imports else ""}
 
 # revision identifiers, used by Alembic.
