--- backend/app/alembic/versions/030be1df4bec_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 030be1df4bec
-Revises: e9da629dde28
-Create Date: 2021-08-27 20:21:41.586282
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '030be1df4bec'
-down_revision = 'e9da629dde28'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
