--- backend/app/alembic/versions/17847892ae1a_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 17847892ae1a
-Revises: 31820b6d5e6c
-Create Date: 2021-08-23 23:37:36.775258
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '17847892ae1a'
-down_revision = '31820b6d5e6c'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
