+++ backend/app/alembic/versions/1c887bb157b0_initial_migration.py
@@ -1,0 +1,91 @@
+"""Initial migration
+
+Revision ID: 1c887bb157b0
+Revises: 
+Create Date: 2021-09-13 16:04:57.920798
+
+"""
+import sqlalchemy as sa
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = '1c887bb157b0'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        'personality', sa.Column('id', sa.Integer(), nullable=False),
+        sa.Column('wikipedia_id', sa.String(), nullable=True),
+        sa.Column('time_created', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
+        sa.Column('field', sa.String(), nullable=True), sa.Column('continent', sa.String(), nullable=True),
+        sa.Column('gender', sa.String(), nullable=True), sa.PrimaryKeyConstraint('id'))
+    op.create_index(op.f('ix_personality_continent'), 'personality', ['continent'], unique=False)
+    op.create_index(op.f('ix_personality_field'), 'personality', ['field'], unique=False)
+    op.create_index(op.f('ix_personality_gender'), 'personality', ['gender'], unique=False)
+    op.create_index(op.f('ix_personality_id'), 'personality', ['id'], unique=False)
+    op.create_index(op.f('ix_personality_wikipedia_id'), 'personality', ['wikipedia_id'], unique=False)
+    op.create_table('user', sa.Column('id', sa.Integer(), nullable=False),
+                    sa.Column('first_name', sa.String(), nullable=True),
+                    sa.Column('last_name', sa.String(), nullable=True), sa.Column('email', sa.String(), nullable=False),
+                    sa.Column('job', sa.String(), nullable=True), sa.Column('organization', sa.String(), nullable=True),
+                    sa.Column('hashed_password', sa.String(), nullable=False),
+                    sa.Column('is_active', sa.Boolean(), nullable=True),
+                    sa.Column('is_superuser', sa.Boolean(), nullable=True), sa.PrimaryKeyConstraint('id'))
+    op.create_index(op.f('ix_user_email'), 'user', ['email'], unique=True)
+    op.create_index(op.f('ix_user_first_name'), 'user', ['first_name'], unique=False)
+    op.create_index(op.f('ix_user_id'), 'user', ['id'], unique=False)
+    op.create_index(op.f('ix_user_job'), 'user', ['job'], unique=False)
+    op.create_index(op.f('ix_user_last_name'), 'user', ['last_name'], unique=False)
+    op.create_index(op.f('ix_user_organization'), 'user', ['organization'], unique=False)
+    op.create_table(
+        'comment', sa.Column('id', sa.Integer(), nullable=False), sa.Column('text', sa.String(), nullable=True),
+        sa.Column('time_created', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
+        sa.Column('fluff', sa.String(), nullable=True), sa.Column('author_id', sa.Integer(), nullable=True),
+        sa.Column('personality_id', sa.Integer(), nullable=True), sa.ForeignKeyConstraint(
+            ['author_id'],
+            ['user.id'],
+        ), sa.ForeignKeyConstraint(
+            ['personality_id'],
+            ['personality.id'],
+        ), sa.PrimaryKeyConstraint('id'))
+    op.create_index(op.f('ix_comment_fluff'), 'comment', ['fluff'], unique=False)
+    op.create_index(op.f('ix_comment_id'), 'comment', ['id'], unique=False)
+    op.create_index(op.f('ix_comment_text'), 'comment', ['text'], unique=False)
+    op.create_table('personal', sa.Column('user_id', sa.Integer(), nullable=False),
+                    sa.Column('personality_id', sa.Integer(), nullable=False),
+                    sa.ForeignKeyConstraint(
+                        ['personality_id'],
+                        ['personality.id'],
+                    ), sa.ForeignKeyConstraint(
+                        ['user_id'],
+                        ['user.id'],
+                    ), sa.PrimaryKeyConstraint('user_id', 'personality_id'))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('personal')
+    op.drop_index(op.f('ix_comment_text'), table_name='comment')
+    op.drop_index(op.f('ix_comment_id'), table_name='comment')
+    op.drop_index(op.f('ix_comment_fluff'), table_name='comment')
+    op.drop_table('comment')
+    op.drop_index(op.f('ix_user_organization'), table_name='user')
+    op.drop_index(op.f('ix_user_last_name'), table_name='user')
+    op.drop_index(op.f('ix_user_job'), table_name='user')
+    op.drop_index(op.f('ix_user_id'), table_name='user')
+    op.drop_index(op.f('ix_user_first_name'), table_name='user')
+    op.drop_index(op.f('ix_user_email'), table_name='user')
+    op.drop_table('user')
+    op.drop_index(op.f('ix_personality_wikipedia_id'), table_name='personality')
+    op.drop_index(op.f('ix_personality_id'), table_name='personality')
+    op.drop_index(op.f('ix_personality_gender'), table_name='personality')
+    op.drop_index(op.f('ix_personality_field'), table_name='personality')
+    op.drop_index(op.f('ix_personality_continent'), table_name='personality')
+    op.drop_table('personality')
+    # ### end Alembic commands ###
