--- backend/app/alembic/versions/2271836fdf5d_back_to_start.py
@@ -1,28 +1,0 @@
-"""back to start
-
-Revision ID: 2271836fdf5d
-Revises: 90bd4117bd1b
-Create Date: 2021-09-09 19:30:20.278704
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '2271836fdf5d'
-down_revision = '90bd4117bd1b'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
