--- backend/app/alembic/versions/31820b6d5e6c_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 31820b6d5e6c
-Revises: a4e558c7b76e
-Create Date: 2021-08-23 21:47:56.232365
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '31820b6d5e6c'
-down_revision = 'a4e558c7b76e'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
