--- backend/app/alembic/versions/31d7452d30c3_comments_fix.py
@@ -1,44 +1,0 @@
-"""Comments fix
-
-Revision ID: 31d7452d30c3
-Revises: 57d342a03232
-Create Date: 2021-08-22 19:37:44.259366
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '31d7452d30c3'
-down_revision = '57d342a03232'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        'comment', sa.Column('id', sa.Integer(), nullable=False), sa.Column('text', sa.String(), nullable=True),
-        sa.Column('time_created', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-        sa.Column('fluff', sa.String(), nullable=True), sa.Column('author_id', sa.Integer(), nullable=True),
-        sa.Column('personality_id', sa.Integer(), nullable=True), sa.ForeignKeyConstraint(
-            ['author_id'],
-            ['user.id'],
-        ), sa.ForeignKeyConstraint(
-            ['personality_id'],
-            ['personality.id'],
-        ), sa.PrimaryKeyConstraint('id'))
-    op.create_index(op.f('ix_comment_fluff'), 'comment', ['fluff'], unique=False)
-    op.create_index(op.f('ix_comment_id'), 'comment', ['id'], unique=False)
-    op.create_index(op.f('ix_comment_text'), 'comment', ['text'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_comment_text'), table_name='comment')
-    op.drop_index(op.f('ix_comment_id'), table_name='comment')
-    op.drop_index(op.f('ix_comment_fluff'), table_name='comment')
-    op.drop_table('comment')
-    # ### end Alembic commands ###
