--- backend/app/alembic/versions/34a0b0847b05_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 34a0b0847b05
-Revises: 17847892ae1a
-Create Date: 2021-08-24 01:04:23.407625
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '34a0b0847b05'
-down_revision = '17847892ae1a'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
