--- backend/app/alembic/versions/37c1a06bf4b2_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 37c1a06bf4b2
-Revises: 840ce790e442
-Create Date: 2021-08-23 01:55:19.500220
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '37c1a06bf4b2'
-down_revision = '840ce790e442'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
