--- backend/app/alembic/versions/3c8221250fca_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 3c8221250fca
-Revises: 7094cbc74bb0
-Create Date: 2021-08-22 23:57:14.283022
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '3c8221250fca'
-down_revision = '7094cbc74bb0'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
