--- backend/app/alembic/versions/40adcb56bd8d_switch_items_to_persons.py
@@ -1,54 +1,0 @@
-"""Switch items to persons
-
-Revision ID: 40adcb56bd8d
-Revises: b5af69df7c8c
-Create Date: 2021-08-16 21:57:45.779263
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '40adcb56bd8d'
-down_revision = 'b5af69df7c8c'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('personality', sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('title', sa.String(), nullable=True), sa.Column('description', sa.String(),
-                                                                              nullable=True),
-                    sa.Column('owner_id', sa.Integer(), nullable=True),
-                    sa.ForeignKeyConstraint(
-                        ['owner_id'],
-                        ['user.id'],
-                    ), sa.PrimaryKeyConstraint('id'))
-    op.create_index(op.f('ix_personality_description'), 'personality', ['description'], unique=False)
-    op.create_index(op.f('ix_personality_id'), 'personality', ['id'], unique=False)
-    op.create_index(op.f('ix_personality_title'), 'personality', ['title'], unique=False)
-    op.drop_index('ix_item_description', table_name='item')
-    op.drop_index('ix_item_id', table_name='item')
-    op.drop_index('ix_item_title', table_name='item')
-    op.drop_table('item')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('item', sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-                    sa.Column('title', sa.VARCHAR(), autoincrement=False, nullable=True),
-                    sa.Column('description', sa.VARCHAR(), autoincrement=False, nullable=True),
-                    sa.Column('owner_id', sa.INTEGER(), autoincrement=False, nullable=True),
-                    sa.ForeignKeyConstraint(['owner_id'], ['user.id'], name='item_owner_id_fkey'),
-                    sa.PrimaryKeyConstraint('id', name='item_pkey'))
-    op.create_index('ix_item_title', 'item', ['title'], unique=False)
-    op.create_index('ix_item_id', 'item', ['id'], unique=False)
-    op.create_index('ix_item_description', 'item', ['description'], unique=False)
-    op.drop_index(op.f('ix_personality_title'), table_name='personality')
-    op.drop_index(op.f('ix_personality_id'), table_name='personality')
-    op.drop_index(op.f('ix_personality_description'), table_name='personality')
-    op.drop_table('personality')
-    # ### end Alembic commands ###
