--- backend/app/alembic/versions/42bad993972d_debugging_time_created.py
@@ -1,28 +1,0 @@
-"""Debugging time_created
-
-Revision ID: 42bad993972d
-Revises: bdc3c703d0cc
-Create Date: 2021-09-09 18:49:42.595204
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '42bad993972d'
-down_revision = 'bdc3c703d0cc'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
