--- backend/app/alembic/versions/46dba5966876_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 46dba5966876
-Revises: 37c1a06bf4b2
-Create Date: 2021-08-23 02:03:59.692234
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '46dba5966876'
-down_revision = '37c1a06bf4b2'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
