--- backend/app/alembic/versions/4edb8fb78b04_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 4edb8fb78b04
-Revises: 97b52bb5c299
-Create Date: 2021-08-27 00:16:18.833418
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '4edb8fb78b04'
-down_revision = '97b52bb5c299'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
