--- backend/app/alembic/versions/57d342a03232_comments.py
@@ -1,30 +1,0 @@
-"""Comments
-
-Revision ID: 57d342a03232
-Revises: 9f1fdf55725b
-Create Date: 2021-08-22 19:34:29.820693
-
-"""
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '57d342a03232'
-down_revision = '9f1fdf55725b'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('personality', 'comments')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('personality',
-                  sa.Column('comments', postgresql.JSON(astext_type=sa.Text()), autoincrement=False, nullable=True))
-    # ### end Alembic commands ###
