--- backend/app/alembic/versions/64f9c5bd1edf_debugging_time_created.py
@@ -1,28 +1,0 @@
-"""Debugging time_created
-
-Revision ID: 64f9c5bd1edf
-Revises: 42bad993972d
-Create Date: 2021-09-09 18:51:56.999005
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '64f9c5bd1edf'
-down_revision = '42bad993972d'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
