--- backend/app/alembic/versions/7094cbc74bb0_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 7094cbc74bb0
-Revises: 71e31dd119d3
-Create Date: 2021-08-22 23:36:15.610124
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '7094cbc74bb0'
-down_revision = '71e31dd119d3'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
