--- backend/app/alembic/versions/71e31dd119d3_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 71e31dd119d3
-Revises: 7d7d1cbed191
-Create Date: 2021-08-22 21:39:47.303635
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '71e31dd119d3'
-down_revision = '7d7d1cbed191'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
