--- backend/app/alembic/versions/71ee2b6e0f6d_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 71ee2b6e0f6d
-Revises: dbb336279d4c
-Create Date: 2021-08-22 19:42:56.712992
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '71ee2b6e0f6d'
-down_revision = 'dbb336279d4c'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
