--- backend/app/alembic/versions/749d2b6f5d70_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 749d2b6f5d70
-Revises: ad25823c6600
-Create Date: 2021-08-24 01:20:29.524279
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '749d2b6f5d70'
-down_revision = 'ad25823c6600'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
