--- backend/app/alembic/versions/76fd9dbebfd0_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 76fd9dbebfd0
-Revises: 46dba5966876
-Create Date: 2021-08-23 02:11:02.170265
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '76fd9dbebfd0'
-down_revision = '46dba5966876'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
