--- backend/app/alembic/versions/7d7d1cbed191_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 7d7d1cbed191
-Revises: f55b3273a916
-Create Date: 2021-08-22 19:46:06.222301
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '7d7d1cbed191'
-down_revision = 'f55b3273a916'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
