--- backend/app/alembic/versions/840ce790e442_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 840ce790e442
-Revises: e8c0633fc1cf
-Create Date: 2021-08-23 01:19:09.269045
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '840ce790e442'
-down_revision = 'e8c0633fc1cf'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
