--- backend/app/alembic/versions/8f8bba24927c_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 8f8bba24927c
-Revises: 2271836fdf5d
-Create Date: 2021-09-10 12:42:18.920682
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '8f8bba24927c'
-down_revision = '2271836fdf5d'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
