--- backend/app/alembic/versions/90bd4117bd1b_reinit_fluff.py
@@ -1,28 +1,0 @@
-"""Reinit fluff
-
-Revision ID: 90bd4117bd1b
-Revises: 64f9c5bd1edf
-Create Date: 2021-09-09 18:53:17.686430
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '90bd4117bd1b'
-down_revision = '64f9c5bd1edf'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
