--- backend/app/alembic/versions/97b52bb5c299_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: 97b52bb5c299
-Revises: de19eb5ca4fc
-Create Date: 2021-08-24 13:00:53.094145
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '97b52bb5c299'
-down_revision = 'de19eb5ca4fc'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
