--- backend/app/alembic/versions/a4e558c7b76e_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: a4e558c7b76e
-Revises: 76fd9dbebfd0
-Create Date: 2021-08-23 20:07:45.238062
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'a4e558c7b76e'
-down_revision = '76fd9dbebfd0'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
