--- backend/app/alembic/versions/ad25823c6600_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: ad25823c6600
-Revises: 34a0b0847b05
-Create Date: 2021-08-24 01:07:57.389380
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'ad25823c6600'
-down_revision = '34a0b0847b05'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
