--- backend/app/alembic/versions/b5af69df7c8c_add_column_last_name_to_user_model.py
@@ -1,50 +1,0 @@
-"""Add column last_name to User model
-
-Revision ID: b5af69df7c8c
-Revises: d4867f3a4c0a
-Create Date: 2021-08-15 01:33:25.554335
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'b5af69df7c8c'
-down_revision = 'd4867f3a4c0a'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('first_name', sa.String(), nullable=True))
-    op.add_column('user', sa.Column('last_name', sa.String(), nullable=True))
-    op.add_column('user', sa.Column('job', sa.String(), nullable=True))
-    op.add_column('user', sa.Column('organization', sa.String(), nullable=True))
-    op.alter_column('user', 'email', existing_type=sa.VARCHAR(), nullable=False)
-    op.alter_column('user', 'hashed_password', existing_type=sa.VARCHAR(), nullable=False)
-    op.drop_index('ix_user_full_name', table_name='user')
-    op.create_index(op.f('ix_user_first_name'), 'user', ['first_name'], unique=False)
-    op.create_index(op.f('ix_user_job'), 'user', ['job'], unique=False)
-    op.create_index(op.f('ix_user_last_name'), 'user', ['last_name'], unique=False)
-    op.create_index(op.f('ix_user_organization'), 'user', ['organization'], unique=False)
-    op.drop_column('user', 'full_name')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('full_name', sa.VARCHAR(), autoincrement=False, nullable=True))
-    op.drop_index(op.f('ix_user_organization'), table_name='user')
-    op.drop_index(op.f('ix_user_last_name'), table_name='user')
-    op.drop_index(op.f('ix_user_job'), table_name='user')
-    op.drop_index(op.f('ix_user_first_name'), table_name='user')
-    op.create_index('ix_user_full_name', 'user', ['full_name'], unique=False)
-    op.alter_column('user', 'hashed_password', existing_type=sa.VARCHAR(), nullable=True)
-    op.alter_column('user', 'email', existing_type=sa.VARCHAR(), nullable=True)
-    op.drop_column('user', 'organization')
-    op.drop_column('user', 'job')
-    op.drop_column('user', 'last_name')
-    op.drop_column('user', 'first_name')
-    # ### end Alembic commands ###
