--- backend/app/alembic/versions/dbb336279d4c_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: dbb336279d4c
-Revises: 31d7452d30c3
-Create Date: 2021-08-22 19:40:22.868266
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'dbb336279d4c'
-down_revision = '31d7452d30c3'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
