--- backend/app/alembic/versions/de19eb5ca4fc_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: de19eb5ca4fc
-Revises: 749d2b6f5d70
-Create Date: 2021-08-24 01:21:02.983059
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'de19eb5ca4fc'
-down_revision = '749d2b6f5d70'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
