--- backend/app/alembic/versions/e8c0633fc1cf_comments_fix.py
@@ -1,36 +1,0 @@
-"""Comments fix
-
-Revision ID: e8c0633fc1cf
-Revises: 3c8221250fca
-Create Date: 2021-08-23 01:17:22.656714
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'e8c0633fc1cf'
-down_revision = '3c8221250fca'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('personal', sa.Column('user_id', sa.Integer(), nullable=False),
-                    sa.Column('personality_id', sa.Integer(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['personality_id'],
-                        ['personality.id'],
-                    ), sa.ForeignKeyConstraint(
-                        ['user_id'],
-                        ['user.id'],
-                    ), sa.PrimaryKeyConstraint('user_id', 'personality_id'))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('personal')
-    # ### end Alembic commands ###
