--- backend/app/alembic/versions/e9da629dde28_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: e9da629dde28
-Revises: 4edb8fb78b04
-Create Date: 2021-08-27 00:36:51.781826
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'e9da629dde28'
-down_revision = '4edb8fb78b04'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
