--- backend/app/alembic/versions/ef8d4c52d6fb_.py
@@ -1,30 +1,0 @@
-"""empty message
-
-Revision ID: ef8d4c52d6fb
-Revises: 8f8bba24927c
-Create Date: 2021-09-11 19:06:16.130403
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'ef8d4c52d6fb'
-down_revision = '8f8bba24927c'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('personality', sa.Column('continent', sa.String(), nullable=True))
-    op.create_index(op.f('ix_personality_continent'), 'personality', ['continent'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_personality_continent'), table_name='personality')
-    op.drop_column('personality', 'continent')
-    # ### end Alembic commands ###
