--- backend/app/alembic/versions/f55b3273a916_comments_fix.py
@@ -1,28 +1,0 @@
-"""Comments fix
-
-Revision ID: f55b3273a916
-Revises: 71ee2b6e0f6d
-Create Date: 2021-08-22 19:44:56.120029
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'f55b3273a916'
-down_revision = '71ee2b6e0f6d'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
